requirements.txt
django
djangorestframework
drf-spectacular
dj-database-url
psycopg[binary]
django-cors-headers
gunicorn
python-dotenv